Rest Table Number data type change in, change for table number data type in reservation repository, little staff model change accordingly with restaurant details

+++ src/main/java/com/wityorestaurant/modules/config/dto/RestTableDTO.java
@@ -1,15 +1,15 @@
 package com.wityorestaurant.modules.config.dto;
 
 public class RestTableDTO {
-    private Long tableNumber;
+    private Integer tableNumber;
     private Integer tableSize;
     private Integer qrCode;
 
-    public Long getTableNumber() {
+    public Integer getTableNumber() {
         return tableNumber;
     }
 
-    public void setTableNumber(Long tableNumber) {
+    public void setTableNumber(Integer tableNumber) {
         this.tableNumber = tableNumber;
     }
